requirements.txt
pytelegrambotapi>=3.8.1
mysql-connector-python>=8.0.25
